Split wasm code from everything else
It was starting to get difficult to track state. So much stuff as
globals, it felt like it was going to turn into spaghetti

Split out an App, index.zig is not just responsible for binding to the
app. Organize stuff in the app in a way that's easier for my brain to
handle