Replace preprocessing with site generator
Our site depends on some outputs of the build, some resource files, and
some generated files. Instead of trying to organize everything at build
time, just make a lil' executable that takes all our inputs and puts
them in the right place

For now we symlink everything to simplify development. In the future we
will probably add a switch to copy instead

Split out an XmlParser as main was starting to feel like it did too much