Work in meters instead of degrees
On app init, take each gps coord and convert it to floating point meters
offset from the center of the map. As a side effect, the GpsCoord type
no longer is correct. Use this as an opportunity to use lin.Point
everywhere, and just have some named aliases for unit conversion. Strong
typing here doesn't feel worth it

The conversion is done in place, so there a few const removals that look
a little odd on first glance