Add rudimentary path planning
Add a button to the UI to start a path plan. On each movement we plan
from the start point to the new center of the screen. This currently
blocks the render loop, but happens relatively quickly, even in the
worst case scenario. Some work has been put into making sure the current
iteration can run fast, but the planner is also set up to support being
polled externally from JS if it comes to that

Initial impl is essentially a copy paste of A* from wikipedia, but
without the fscore tracking array as a speed optimization.

Supporting changes...
* getNeighbors() call changed to avoid complexity in getNeighbors().
  Turns out spending 4 extra bytes of memory is noticeably worth the
  reduction in complexity
* Finally support RGB in our fragment shader to draw the path in a
  different color than the selected way
* Stash closest node on mouse move so that we can start a path plan
* renderPoints updated to allow rendering a line out of a bunch of
  points with glDrawElements
  * allocate an ebo on start for working with ephemeral data
* Node processing in make_site updated
  * Only write nodes if they are used in a way
  * Add nodes in order that they appear in ways. This has massive CPU
    caching advantages and ~3xs path planner performance