Add momentum to path planner
Previously, in all planned paths we picked the shortest path. This is
annoying if the shortest path contains a lot of turns. I'd happily walk
a little farther, if it was significantly easier to remember where to go

Add concept of turning cost. This adds a constant penalty equivalent to
extra distance traveled for every turn. This seems to fix the zig zag
problem at the cost of a much larger graph to traverse

In practice the algo is still "fast enough" so commit as is, improve
later if necessary

Represent graph as pairs of nodes instead of individual nodes. The
rationale here is that the cost of moving from B -> C depends on the
movement from A -> B. Therefore we need A->B->C for each edge
traversal. This essentially adds a dimension to the graph. Think of
each node as a set of nodes containing where we came from

Bonus changes:
* Log average path planning time in benchmark program
* Do not crash on mouse move, just log an error